rename Sentry to TSentry
windows filenames arent case-sensitive so it doesnt know which
sentry.h we mean.
